backend/requirements.txt
# Core FastAPI and web framework - latest compatible versions
fastapi>=0.104.0,<0.117.0
uvicorn[standard]>=0.24.0,<0.36.0
pydantic>=2.5.0,<3.0.0
starlette>=0.40.0,<0.49.0
python-multipart>=0.0.6

# HTTP clients and requests - compatible versions
httpx>=0.25.0,<0.28.0
requests>=2.31.0,<2.33.0
urllib3>=2.0.0,<3.0.0

# Environment and configuration
python-dotenv>=1.0.0

# AI/ML APIs - latest versions
anthropic>=0.25.0
groq>=0.8.0

# Data processing - lighter requirements for deployment
pandas>=2.0.0,<2.3.0
numpy>=1.24.0,<2.0.0

# Optional ML packages - only if needed
# scikit-learn>=1.3.0,<1.6.0
# sentence-transformers>=2.2.0,<3.0.0
# transformers>=4.30.0,<5.0.0

# Optional visualization - only if needed  
# networkx>=3.0.0,<4.0.0
# matplotlib>=3.7.0,<4.0.0
# plotly>=5.15.0,<6.0.0

# Date/time handling
python-dateutil>=2.8.0,<3.0.0